+++ myosuite/envs/myo/key_turn_v0.py
@@ -22,23 +22,21 @@ class KeyTurnEnvV0(BaseV0):
         'penalty':25.0
     }
 
-    def __init__(self,
-                model_path:str,
-                **kwargs):
+    def __init__(self, model_path, obsd_model_path=None, seed=None, **kwargs):
 
         # EzPickle.__init__(**locals()) is capturing the input dictionary of the init method of this class.
         # In order to successfully capture all arguments we need to call gym.utils.EzPickle.__init__(**locals())
         # at the leaf level, when we do inheritance like we do here.
         # kwargs is needed at the top level to account for injection of __class__ keyword.
         # Also see: https://github.com/openai/gym/pull/1497
-        gym.utils.EzPickle.__init__(**locals())
+        gym.utils.EzPickle.__init__(self, model_path, obsd_model_path, seed, **kwargs)
 
         # This two step construction is required for pickling to work correctly. All arguments to all __init__
         # calls must be pickle friendly. Things like sim / sim_obsd are NOT pickle friendly. Therefore we
         # first construct the inheritance chain, which is just __init__ calls all the way down, with env_base
         # creating the sim / sim_obsd instances. Next we run through "setup"  which relies on sim / sim_obsd
         # created in __init__ to complete the setup.
-        super().__init__(model_path=model_path)
+        super().__init__(model_path=model_path, obsd_model_path=obsd_model_path, seed=seed)
 
         self._setup(**kwargs)
 
@@ -54,6 +52,7 @@ class KeyTurnEnvV0(BaseV0):
         self.IF_sid = self.sim.model.site_name2id("IFtip")
         self.TH_sid = self.sim.model.site_name2id("THtip")
         self.key_init_range = key_init_range
+        self.key_init_pos = self.sim.data.site_xpos[self.keyhead_sid].copy()
 
         super()._setup(obs_keys=obs_keys,
                     weighted_reward_keys=weighted_reward_keys,
@@ -116,6 +115,6 @@ class KeyTurnEnvV0(BaseV0):
         qvel = self.init_qvel.copy() if reset_qvel is None else reset_qvel
         qpos[-1] = self.np_random.uniform(low=self.key_init_range[0], high=self.key_init_range[1])
         if self.key_init_range[0]!=self.key_init_range[1]: # randomEnv
-            self.sim.model.body_pos[-1] = np.array([-.15, -.23, 1.025])+self.np_random.uniform(low=np.array([-0.01, -0.01, -.01]), high=np.array([0.01, 0.01, 0.01]))
+            self.sim.model.body_pos[-1] = self.key_init_pos+self.np_random.uniform(low=np.array([-0.01, -0.01, -.01]), high=np.array([0.01, 0.01, 0.01]))
         self.robot.reset(qpos, qvel)
         return self.get_obs()