+++ myosuite/utils/quat_math.py
@@ -41,7 +41,7 @@ def axis_angle2quat(axis, angle):
     return np.array([c, s*axis[0], s*axis[1], s*axis[2]])
 
 def euler2mat(euler):
-    """ Convert Euler Angles to Rotation Matrix.  See rotation.py for notes """
+    """ Convert Euler Angles to Rotation Matrix """
     euler = np.asarray(euler, dtype=np.float64)
     assert euler.shape[-1] == 3, "Invalid shaped euler {}".format(euler)
 
@@ -65,7 +65,7 @@ def euler2mat(euler):
 
 
 def euler2quat(euler):
-    """ Convert Euler Angles to Quaternions.  See rotation.py for notes """
+    """ Convert Euler Angles to Quaternions """
     euler = np.asarray(euler, dtype=np.float64)
     assert euler.shape[-1] == 3, "Invalid shape euler {}".format(euler)
 
@@ -84,7 +84,7 @@ def euler2quat(euler):
 
 
 def mat2euler(mat):
-    """ Convert Rotation Matrix to Euler Angles.  See rotation.py for notes """
+    """ Convert Rotation Matrix to Euler Angles """
     mat = np.asarray(mat, dtype=np.float64)
     assert mat.shape[-2:] == (3, 3), "Invalid shape matrix {}".format(mat)
 
@@ -104,7 +104,7 @@ def mat2euler(mat):
 
 
 def mat2quat(mat):
-    """ Convert Rotation Matrix to Quaternion.  See rotation.py for notes """
+    """ Convert Rotation Matrix to Quaternion """
     mat = np.asarray(mat, dtype=np.float64)
     assert mat.shape[-2:] == (3, 3), "Invalid shape matrix {}".format(mat)
 
@@ -141,12 +141,12 @@ def mat2quat(mat):
 
 
 def quat2euler(quat):
-    """ Convert Quaternion to Euler Angles.  See rotation.py for notes """
+    """ Convert Quaternion to Euler Angles """
     return mat2euler(quat2mat(quat))
 
 
 def quat2mat(quat):
-    """ Convert Quaternion to Euler Angles.  See rotation.py for notes """
+    """ Convert Quaternion to Euler Angles """
     quat = np.asarray(quat, dtype=np.float64)
     assert quat.shape[-1] == 4, "Invalid shape quat {}".format(quat)