+++ .github/CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1] - 2022-10-09
+- update horizon for MyoChallenge Die Reorient task
+- update tutorials
+
+## [1.2.0] - 2022-08-13
+- Rebase and building on RoboHive v0.2
+- Adding the myochallenge envs
+- Fundamental bugfixes on the RoboHive engine
+- Bugfixes on myo environments as well
+- Closes baselines are on RoboHive-v0.2
+- Next planned baseline release will align when Robohive-v0.3dev moves to prerelease.
+- Renaming the metrics for clarity and changed sign from `act_mag` to `effort` and `solved` to `score`
+
+## [1.1.0] - 2022-08-12
+- Upgrade to mj_env v0.2 experimental
+- add Die Rotation and Baoding Ball task for MyoChallenge (https://sites.google.com/view/myochallenge)
+
 ## [1.0.1] - 2022-05-23
 - First Release of MyoSuite.
 - Basic Documentation